start to handle a length of resources
